# Simpler (faster?) true/false/null literal checking