This patch simplifies, and possibly speeds up, parsing of the literals `true`, `false`, and `null`. The current benchmarks do not test the parsing of booleans or null, so it's hard to say for sure whether the speed is improved.

**Benchmark after:**
BenchmarkJsonParserLarge-8 	  100000	     56923 ns/op
BenchmarkJsonParserMedium-8	  500000	     11938 ns/op
BenchmarkJsonParserSmall-8 	10000000	       915 ns/op

**Benchmark before:**
BenchmarkJsonParserLarge-8 	  200000	     56920 ns/op
BenchmarkJsonParserMedium-8	 1000000	     11469 ns/op
BenchmarkJsonParserSmall-8 	10000000	       965 ns/op