@buger 2016-03-30
Less magic is always better, merged